% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadox-package.R
\docType{package}
\name{datadox-package}
\alias{datadox}
\alias{datadox-package}
\title{datadox: Package Data Documentation}
\description{
Helpers to create data dictionaries and use them to document package datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wranglezone/datadox}
  \item \url{https://wranglezone.github.io/datadox/}
  \item Report bugs at \url{https://github.com/wranglezone/datadox/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
