% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawrap-package.R
\docType{package}
\name{datawrap-package}
\alias{datawrap}
\alias{datawrap-package}
\title{datawrap: Final Steps for Dataset Preparation}
\description{
Helpers to create data dictionaries, document package datasets, and apply finishing touches to data for distribution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wranglezone/datawrap}
  \item \url{https://wranglezone.github.io/datawrap/}
  \item Report bugs at \url{https://github.com/wranglezone/datawrap/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
