% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared_params.R
\name{.shared-params}
\alias{.shared-params}
\title{Shared parameters}
\arguments{
\item{call}{(\code{environment}) The caller environment for error messages.}

\item{dataset}{(\code{data.frame}) The dataset to document.}

\item{open}{(\code{logical(1)}) Whether to open the file for editing after
writing. Defaults to \code{TRUE} in interactive sessions.}
}
\description{
These parameters are used in multiple functions. They are defined here to
make them easier to import and to find.
}
\keyword{internal}
