% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset_dictionary.R
\name{.str_to_sentence_full}
\alias{.str_to_sentence_full}
\title{Convert a string to a full sentence}
\usage{
.str_to_sentence_full(string, locale = "en")
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{locale}{Locale to use for comparisons. See
\code{\link[stringi:stri_locale_list]{stringi::stri_locale_list()}} for all possible options.
Defaults to "en" (English) to ensure that default behaviour is
consistent across platforms.}
}
\value{
(\code{character}) The input string(s) converted to sentence case with a
trailing period.
}
\description{
Convert a string to a full sentence
}
\keyword{internal}
