% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataset_dictionary.R
\name{.write_file_lines}
\alias{.write_file_lines}
\title{Write lines to a file and optionally open it for editing}
\usage{
.write_file_lines(lines, file_path, open = rlang::is_interactive())
}
\arguments{
\item{lines}{(\code{character}) The lines to write.}

\item{file_path}{(\code{character(1)}) Path to the output file.}

\item{open}{(\code{logical(1)}) Whether to open the file for editing after
writing. Defaults to \code{TRUE} in interactive sessions.}
}
\value{
(\code{character(1)}) The path to the written file, invisibly.
}
\description{
Write lines to a file and optionally open it for editing
}
\keyword{internal}
